refactor: deduplicate buildModelConfig into shared util (closes #21)
Extract the buildModelConfig function (copy-pasted in serve.ts, route.ts,
and daemon/index.ts) into src/utils/model-config.ts and import from there.
