+++ src/cli/commands/route.ts
@@ -7,6 +7,7 @@ import { getDb, initSchema } from '../../db/client.js'
 import { selectModel } from '../../router/selector.js'
 import { callModel } from '../../providers/compat.js'
 import type { ModelConfig } from '../../types/index.js'
+import { buildModelConfig } from '../../utils/model-config.js'
 
 interface RouteCommandOpts {
   type?: string
@@ -104,44 +105,3 @@ export async function routeCommand(prompt: string, opts: RouteCommandOpts): Prom
   }
 }
 
-/** Build a minimal ModelConfig from model ID and provider. */
-function buildModelConfig(modelId: string, provider: string): ModelConfig {
-  const config: ModelConfig = {
-    id: modelId,
-    model: modelId,
-    api_key: 'none',
-    tags: [],
-    port: 8080,
-    timeout_ms: 120_000,
-    max_tokens: 2048,
-  }
-
-  switch (provider) {
-    case 'ollama': {
-      const host = process.env.OLLAMA_HOST ?? 'http://localhost:11434'
-      config.base_url = host.endsWith('/v1') ? host : `${host}/v1`
-      break
-    }
-    case 'mlx': {
-      const port = process.env.MLX_PORT ?? '8080'
-      config.base_url = `http://localhost:${port}/v1`
-      break
-    }
-    case 'openrouter':
-      config.base_url = 'https://openrouter.ai/api/v1'
-      config.api_key = process.env.OPENROUTER_API_KEY ?? 'none'
-      break
-    case 'openai':
-      config.base_url = 'https://api.openai.com/v1'
-      config.api_key = process.env.OPENAI_API_KEY ?? 'none'
-      break
-    default: {
-      // Try ollama as default for local models
-      const host = process.env.OLLAMA_HOST ?? 'http://localhost:11434'
-      config.base_url = host.endsWith('/v1') ? host : `${host}/v1`
-      break
-    }
-  }
-
-  return config
-}
